+++ HomeWork_16/testNGPositive.xml
@@ -1,0 +1,10 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite name="HW_16_Positive_Suite">
+    <test name="HW_16_Positive_Test">
+        <classes>
+            <class name="com.example.MyTestClass1"/>
+            <class name="com.example.MyTestClass2"/>
+            <!-- Добавьте сюда другие классы тестов -->
+        </classes>
+    </test>
+</suite>
